api/pyproject.toml
[project]
name = "mas-v0.2"
version = "0.1.0"
description = "Multi-Agent System for Digital Marketing"
authors = [
    {name = "HuanMu Team", email = "consumerclone@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "alembic>=1.16.4",
    "anthropic>=0.64.0",
    "asyncpg>=0.30.0",
    "colorlog>=6.9.0",
    "cryptography>=45.0.6",
    "elasticsearch>=9.1.0",
    "fastapi>=0.116.1",
    "google-genai>=1.30.0",
    "greenlet>=3.2.4",
    "langchain>=0.3.27",
    "langfuse>=3.3.1",
    "langgraph>=0.6.5",
    "msgpack>=1.1.1",
    "openai>=1.100.0",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymilvus>=2.6.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
